# public/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Visor TV Enterprise — API RESTful",
    "description": "Especificación formal de la API de Señalización Digital y Transmisión de Contenidos para Smart TVs y pantallas comerciales.\n\n**Desarrollado y Diseñado por Ashly Nicole**\n\n* Autenticación mediante Bearer Token (Sanctum)\n* Streaming HTTP 206 Partial Content (Range requests) y soporte HEAD\n* Sincronización dinámica Zero-Delay mediante firma MD5 de versión\n* Auditoría completa de eventos en tiempo real",
    "version": "2.0.0",
    "contact": {
      "name": "Ashly Nicole — Lead Engineer",
      "email": "ashly@visortv.com"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Servidor Local de Desarrollo"
    },
    {
      "url": "/",
      "description": "Servidor Relativo / Producción"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Sanctum Token",
        "description": "Ingresa el token obtenido en /api/auth/login"
      }
    },
...
  },
  "paths": {
...
    "/api/media/stream/{filename}": {
      "get": {
        "summary": "Streaming HTTP 206 / HEAD",
        "description": "Transmite el video o imagen en bloques de 512 KB con soporte de cabeceras Range, ETag y HEAD para Smart TVs.",
        "parameters": [
          { "name": "filename", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "thumb", "in": "query", "schema": { "type": "integer", "enum": [0, 1] } }
        ],
        "responses": {
          "200": { "description": "Contenido completo" },
          "206": { "description": "Fragmento solicitado (Partial Content)" },
          "304": { "description": "No modificado (Caché válida)" }
        }
      }
    },
...
    "/api/media": {
      "get": {
        "summary": "Listar Medios de Sede",
        "parameters": [
          { "name": "sede_id", "in": "query", "required": true, "schema": { "type": "integer" } },
          { "name": "active_only", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Lista de medios" } }
      },
      "post": {
        "summary": "Subir Archivo Multimedia",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["sede_id", "file"],
                "properties": {
                  "sede_id": { "type": "integer" },
                  "file": { "type": "string", "format": "binary" },
                  "title": { "type": "string" },
                  "duration": { "type": "integer", "default": 10 },
                  "fit_mode": { "type": "string", "enum": ["contain", "cover", "fill"] }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Archivo cargado" } }
      }
    },
...
  }
}